Add option resolve_symlinks for shared collections
To make public collections visible for everybody else
one can use a hack and symlink all directories that a user
has access to to the user collection. To actually get the right
permissions on these files, the symlinks have to be followed to the
actual radicale path. The option resolve_symlinks does just that.